Move RuleSet class to libs/domain/rule.py and update import paths for improved organization
