+++ cls/config.py
@@ -11,8 +11,8 @@ from pathlib import Path, PosixPath
 from types import NoneType
 from typing import TYPE_CHECKING, Any, Literal, Optional, TypeAlias, Union
 
-from cls.rule import RuleSet
 from libs.data.lookup import read_memberslist
+from libs.domain.rule import RuleSet
 from libs.types import GradeTableDict
 
 if TYPE_CHECKING:
