+++ libs/domain/rule.py
